+++ UCR.Tests/UtilityTests/HelperTests/SensitivityHelperTests.cs
@@ -10,7 +10,7 @@ namespace HidWizards.UCR.Tests.UtilityTests.HelperTests
         [TestCase(Constants.AxisMaxValue, ExpectedResult = Constants.AxisMaxValue, TestName = "SensitivityHelper (Init): Max returns Max")]
         [TestCase(Constants.AxisMinValue, ExpectedResult = Constants.AxisMinValue, TestName = "SensitivityHelper (Init): Min returns Min")]
         [TestCase(0, ExpectedResult = 0, TestName = "SensitivityHelper (Init): 0 returns 0")]
-        public long SensitivityHelperInitTests(long inputValue)
+        public short SensitivityHelperInitTests(short inputValue)
         {
             var helper = new SensitivityHelper();
             return helper.ApplyRangeSensitivity(inputValue);
@@ -22,7 +22,7 @@ namespace HidWizards.UCR.Tests.UtilityTests.HelperTests
         [TestCase(Constants.AxisMinValue, 20, ExpectedResult = Constants.AxisMinValue, TestName = "SensitivityHelper (20): Min returns Min")]
         [TestCase(Constants.AxisMaxValue, 200, ExpectedResult = Constants.AxisMaxValue, TestName = "SensitivityHelper (200): Min returns Min")]
         [TestCase(Constants.AxisMinValue, 200, ExpectedResult = Constants.AxisMinValue, TestName = "SensitivityHelper (200): Min returns Min")]
-        public long SensitivityHelperValueTests(long inputValue, int percentage)
+        public short SensitivityHelperValueTests(short inputValue, int percentage)
         {
             var helper = new SensitivityHelper {Percentage = percentage};
             return helper.ApplyRangeSensitivity(inputValue);
@@ -36,7 +36,7 @@ namespace HidWizards.UCR.Tests.UtilityTests.HelperTests
         [TestCase(Constants.AxisMinValue, 200, ExpectedResult = Constants.AxisMinValue, TestName = "SensitivityHelper (200, Linear): Min returns Min")]
         [TestCase(16384, 200, ExpectedResult = Constants.AxisMaxValue, TestName = "SensitivityHelper (200, Linear): Half (Positive) Deflection returns Max")]
         [TestCase(-16384, 200, ExpectedResult = Constants.AxisMinValue, TestName = "SensitivityHelper (200, Linear): Half (Negative) Deflection returns Min")]
-        public long SensitivityHelperValueLinearTests(long inputValue, int percentage)
+        public short SensitivityHelperValueLinearTests(short inputValue, int percentage)
         {
             var helper = new SensitivityHelper { Percentage = percentage, IsLinear = true};
             return helper.ApplyRangeSensitivity(inputValue);
